+++ public/index.html
@@ -35,7 +35,7 @@
     </section>
   </footer>
 
-  <script src="../src/main.js"></script>
+  <script src="./bundle.js"></script>
 </body>
 
 </html>
