+++ src/components/buttonLoadMore.js
@@ -1,0 +1,11 @@
+/**
+ * Создаем разметку Кнопка «Show more»
+ * @return {void}
+ */
+const createButtonLoadMore = () => {
+  return (
+    `<button class="films-list__show-more">Show more</button>`
+  );
+};
+
+export {createButtonLoadMore};
